pyproject.toml
[tool.poetry]
name = "logs-collection"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.0.0"
djangorestframework = "^3.15.0"
django-compat = "^1.0.15"
apscheduler = "^3.10.4"
drf-spectacular = "^0.27.1"
django-rest-enumfield = "^0.2.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.4.1"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
